+++ alchemist-loading/src/main/kotlin/it/unibo/alchemist/boundary/LoadAlchemist.kt
@@ -10,7 +10,7 @@
 package it.unibo.alchemist.boundary
 
 import it.unibo.alchemist.boundary.loader.SimulationModel
-import it.unibo.alchemist.model.VariablesOverrider
+import it.unibo.alchemist.model.util.VariablesOverrider
 import it.unibo.alchemist.util.ClassPathScanner
 import java.io.File
 import java.io.InputStream
