+++ alchemist-loading/src/main/kotlin/it/unibo/alchemist/model/util/VariablesOverrider.kt
@@ -7,7 +7,7 @@
  * as described in the file LICENSE in the Alchemist distribution's top directory.
  */
 
-package it.unibo.alchemist.model
+package it.unibo.alchemist.model.util
 
 import it.unibo.alchemist.boundary.modelproviders.YamlProvider
